requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
coverage==6.5.0
dj3-cloudinary-storage==0.0.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
flake8==5.0.4
idna==3.4
mccabe==0.7.0
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
